Commit for tfarina
Original CL:
http://codereview.chromium.org/155060
This patch changes the behavior of clear search button in cookies_view to work
as in firefox.
When the dialog is launched the clear search button in the old cookie_view the
state of this button is enabled,
but the text field is empty, so the correct behavior is  make the button
disabled. Only when the user start typing
in the text field, the clear button will be enabled.

BUG=None
TEST=None



git-svn-id: svn://svn.chromium.org/chrome/trunk/src@20470 0039d316-1c4b-4281-b951-d872f2087c98