The compiled version of the o3djs libs WORK!
Sadly there is some really strange voodoo to make it work.
util.js was using document.getElementsByTagName('script')
and it was not returning all the scripts. Lots of dump()
lines later the voodoo of calling 
document.getElementsByTagName('script').length (the length
is important) in some place earlier in the code fixes the 
issue.

Also, added the copyright to the compiled file

Added parameter docs


Review URL: http://codereview.chromium.org/159049

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21085 0039d316-1c4b-4281-b951-d872f2087c98