Info bubble focus tweaks.
We can't close info bubbles when they lose focus because then we can't use the select dropdown in the bookmark bubble. Thus we have to watch the active window. Since we are doing this, we no longer need the application level grab or the button press handler.

This also fixes the problem that a browser window showing an info bubble will paint inactive (assuming your WM supports active/inactive windows).

This maintains the desirable properties:
- info bubbles close when you click outside them but inside the parent window
- info bubbles close when the parent window closes
- info bubbles don't show up on alternate desktops (even if your WM doesn't support active/inactive windows)
- info bubbles can survive losing focus to a popup window

This breaks the desirable properties:
- clicking on something in the browser while the info bubble is showing both dismisses the info bubble and does the appropriate action in the browser
- the info bubble dismisses when the browser loses focus in a window manager that doesn't support active/inactive state
- on xmonad, if you have the cursor over the browser window but the focus is on the info bubble (this is hard to do, since focus is supposed to follow the mouse, but not impossible), then it requires two clicks to dismiss the info bubble. But I think that this is a problem with xmonad and is present before this patch as well.

So this is not the best of all possible worlds, but it is an improvement over our current state as well as an improvement over the state before r20977.

BUG=17223
TEST=browser doesn't crash when you open the select dropdown in the bookmark bubble

Review URL: http://codereview.chromium.org/155793

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21141 0039d316-1c4b-4281-b951-d872f2087c98