Fix another d&d crash that happens when the tab contents is swapped out while a drag is happening.
Since the tab contents is swapped out, our RVH that the drag
started on is no longer valid.  Simply null out the pointer
and no longer try to send feedback to the drag source.  This
allows the user to continue the drag even though the source
RVH is gone.

BUG=16073

Review URL: http://codereview.chromium.org/159040

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21187 0039d316-1c4b-4281-b951-d872f2087c98