python 2.4's implementation of sys.popen() has a race condition
python 2.4's implementation of sys.popen() keeps a list of subprocesses
around to handle cleanup, but the list isn't safe against concurrent access,
leading it to sometimes through a ValueError exception. The list is kind
of pointless, and the exception can be ignored (which it is in 2.5).

BUG=none
R=maruel
TEST=none

Review URL: http://codereview.chromium.org/155782

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21367 0039d316-1c4b-4281-b951-d872f2087c98