Prep for moving Camera Info out of Params into JSON
I thought it was best to give you what I have so far
for your input and code review instead of giving you
everything all at once.

This CL implements JSONObject which has a mechanism
for adding stuff to be serialized as JSON to the
serialization code. Since JSONObject is a serialization
only object it seemed okay to put Serialize code
inside.

I opted to pre-declare JSONFloat and JSONOptionalFloat
because I felt it made the code more error free.
I had made it were you could just have JSONFloat and
then with RegisterJSONValue you'd pass in optional
or not but this way, declaring the field in a class
makes it more explicit.

CameraInfo is the first class that uses it. I'm not
set on exactly how it is serialized. Whether it's
as "object: { ... }" inside the "properties" section
or whether it should have its own section. I think
I won't know what until I actually write the 
deserialization code. That's not imporant for this CL

This code, even if checked it, is not used yet
as the import code, collada.cc, is not yet creating
any of these objects. That's in another CL if you want
to take a look

http://codereview.chromium.org/160007

That CL will have to have lots of o3djs changes,
corresponding sample changes and corresponding
selenium test changes.


Review URL: http://codereview.chromium.org/160008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@21515 0039d316-1c4b-4281-b951-d872f2087c98