Commit patch for Pierre-Antoine LaFayette, original CL: http://codereview.chromium.org/164119
Items in the Downloads page should be loaded like a normal OS file when dragged
to a Tab content area. We need the URL to be set to the file path in the
IDataObject so that WebDropData can properly handle the file.

BUG=9266
TEST=Drag an item with a supported MIME type from the Downloads page to a Tab
content area. It should load within the browser. Drag an item with an
unsupported MIME type and it should not open it.

Review URL: http://codereview.chromium.org/165260

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@22983 0039d316-1c4b-4281-b951-d872f2087c98