Remove WebView::SetDelegate because I'd like to avoid having a method like this in the WebKit API.
The only consumer was TestShell.  It was using this method to replace its
TestWebViewDelegate instance.  Instead, with this change, it has a manual Reset
method.  To avoid duplication with the constructor, Reset uses operator=().

This required a couple changes:

1- Remove DISALLOW_COPY_AND_ASSIGN from WebViewDelegate.  Anyways, that didn't
make sense since you cannot 'copy' a class with pure virtual methods.

2- Change scoped_ptr members of TestWebViewDelegate to linked_ptr.  The extra
overhead of the linked_ptr seems warranted in this case.

I also changed TestWebViewDelegate to not be reference counted since it wasn't
necessary.

R=tony
BUG=none
TEST=none

Review URL: http://codereview.chromium.org/164308

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23068 0039d316-1c4b-4281-b951-d872f2087c98