Change selenium to run the samples from the artifacts directory.
Note: There is the issue that the scons build does not
have the correct dependencies to make sure the samples
get copied to the artifacts before selenium runs.

If that's a 1 line fix then let's add it. Otherwise
we can fix that in the gyp build?

Review URL: http://codereview.chromium.org/172008

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23565 0039d316-1c4b-4281-b951-d872f2087c98