Linux: Revert updating PRIMARY selection on Ctrl-C in omnibox.
This backs out r23596.  I'm pretty sure that it worked when I tested it
right before submitting, but it's totally breaking stuff now -- it looks
like when I update the PRIMARY selection, GTK notices that it's lost the
selection and automatically unhighlights the text in the omnibox, making
it impossible to use Ctrl-C to copy omnibox text to the clipboard. :-(

BUG=19648
TESTED=Hit Ctrl-L, Ctrl-C and confirmed that text remains highlighted and becomes CLIPBOARD selection.

Review URL: http://codereview.chromium.org/174058

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@23714 0039d316-1c4b-4281-b951-d872f2087c98