Edit Search Engine fix up:
Fix the class name to match a rename that happened before commit.
Make the label size changes force the fields over, but not string the fields (window grows).
Make the url description autosize to the text it gets.
Make the window track the height needed for the url description field.

BUG=21893
TEST= The url description text should stay under the url box in all languages.  The labels should never clip or overlap the text fields.
Review URL: http://codereview.chromium.org/195096

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@26342 0039d316-1c4b-4281-b951-d872f2087c98