Fix window restore behaviour on Mac.
On the mac, there can be no windows, but the application is sitll open. This
confused "Restore the pages that were open last" so that it would restore
when opening a new window.

BUG=13341
TEST=Open chromium. Set On Startup to "Restore the pages that were open last".
Open a few tabs. Close the window (but not chromium). Click the dock icon.
The tabs shouldn't be restored. But if you quit chrome and start again,
things should reopen.

Review URL: http://codereview.chromium.org/340023

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@30463 0039d316-1c4b-4281-b951-d872f2087c98